+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 * TODO: Describe initial release.
 There are three addintional widgets added to use.
 ## Main Changes in V0.0.4
-Image picker version updated. The icon for Image Picker now flexible. You can move it up or down. The value between 0-100.
+Image picker version updated. The icon for Image Picker now flexible. You can move it up or down. The value between 0-100. You're now able to use our buildin image picker method