+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-## 0.0.5
+## 0.0.6
 
 * TODO: Describe initial release.
 There are three addintional widgets added to use.
-## Main Changes in V0.0.5
-Image picker version updated. The icon for Image Picker now flexible. You can move it up or down. The value between 0-100. You're now able to use our buildin image picker method. You can pass the string image url if needed
+## Main Changes in V0.0.6
+Image picker version updated. The icon for Image Picker now flexible. You can move it up or down. The value between 0-100. You're now able to use our buildin image picker method. You can pass the string image url if needed.